Tests + cleanup
