+++ pkg/logs/config/integration_config.go
@@ -23,7 +23,7 @@ const (
 	// UTF16BE for UTF-16 Big endian encoding
 	UTF16BE string = "utf-16-be"
 	// UTF16LE for UTF-16 Little Endian encoding
-	UTF16LE string = "utf-16-be"
+	UTF16LE string = "utf-16-le"
 )
 
 // LogsConfig represents a log source config, which can be for instance
