+++ pkg/logs/parser/parser_test.go
@@ -18,3 +18,43 @@ func TestNoopParserHandleMessages(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Equal(t, testMsg, msg)
 }
+
+func TestUTF16LEParserHandleMessages(t *testing.T) {
+	parser := NewDecodingParser(UTF16LE)
+	testMsg := []byte{'F', 0x0, 'o', 0x0, 'o', 0x0}
+	msg, _, _, err := parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+
+	// We should support BOM
+	testMsg = []byte{0xFF, 0xFE, 'F', 0x0, 'o', 0x0, 'o', 0x0}
+	msg, _, _, err = parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+
+	// BOM overrides endianness
+	testMsg = []byte{0xFE, 0xFF, 0x0, 'F', 0x0, 'o', 0x0, 'o'}
+	msg, _, _, err = parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+}
+
+func TestUTF16BEParserHandleMessages(t *testing.T) {
+	parser := NewDecodingParser(UTF16BE)
+	testMsg := []byte{0x0, 'F', 0x0, 'o', 0x0, 'o'}
+	msg, _, _, err := parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+
+	// We should support BOM
+	testMsg = []byte{0xFE, 0xFF, 0x0, 'F', 0x0, 'o', 0x0, 'o'}
+	msg, _, _, err = parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+
+	// BOM overrides endianness
+	testMsg = []byte{0xFF, 0xFE, 'F', 0x0, 'o', 0x0, 'o', 0x0}
+	msg, _, _, err = parser.Parse(testMsg)
+	assert.Nil(t, err)
+	assert.Equal(t, "Foo", string(msg))
+}
